+++ .idea/misc.xml
@@ -9,6 +9,7 @@
         <entry key="..\:/Users/Phu Tuan/AndroidStudioProjects/MyApplication3/app/src/main/res/layouts/LOGIN/layout/activity_main.xml" value="0.18802083333333333" />
         <entry key="..\:/Users/Phu Tuan/AndroidStudioProjects/MyApplication3/app/src/main/res/layouts/LOGIN/layout/forgetpass.xml" value="0.19895833333333332" />
         <entry key="..\:/Users/Phu Tuan/AndroidStudioProjects/MyApplication3/app/src/main/res/layouts/LOGIN/layout/forgetpass1.xml" value="0.19895833333333332" />
+        <entry key="..\:/Users/Phu Tuan/AndroidStudioProjects/MyApplication3/app/src/main/res/layouts/LOGIN/layout/forgetpass2.xml" value="0.18802083333333333" />
         <entry key="..\:/Users/Phu Tuan/AndroidStudioProjects/MyApplication3/app/src/main/res/layouts/LOGIN/layout/forgetpass3.xml" value="0.19895833333333332" />
         <entry key="..\:/Users/Phu Tuan/AndroidStudioProjects/MyApplication3/app/src/main/res/layouts/LOGIN/layout/login1.xml" value="0.18802083333333333" />
         <entry key="..\:/Users/Phu Tuan/AndroidStudioProjects/MyApplication3/app/src/main/res/layouts/LOGIN/layout/login2.xml" value="0.18802083333333333" />
@@ -23,6 +24,7 @@
         <entry key="..\:/Users/Phu Tuan/AndroidStudioProjects/MyApplication3/app/src/main/res/layouts/SV/layout/svdiemdanhqr.xml" value="0.358695652173913" />
         <entry key="..\:/Users/Phu Tuan/AndroidStudioProjects/MyApplication3/app/src/main/res/layouts/SV/layout/svdkmon.xml" value="0.358695652173913" />
         <entry key="..\:/Users/Phu Tuan/AndroidStudioProjects/MyApplication3/app/src/main/res/layouts/SV/layout/svinfo.xml" value="0.358695652173913" />
+        <entry key="..\:/Users/Phu Tuan/AndroidStudioProjects/MyApplication3/app/src/main/res/layouts/SV/layout/svinfochange.xml" value="0.18802083333333333" />
         <entry key="..\:/Users/Phu Tuan/AndroidStudioProjects/MyApplication3/app/src/main/res/layouts/SV/layout/svlichhoc.xml" value="0.358695652173913" />
         <entry key="..\:/Users/Phu Tuan/AndroidStudioProjects/MyApplication3/app/src/main/res/layouts/SV/layout/svpasschange.xml" value="0.358695652173913" />
       </map>
