+++ app/src/main/java/com/example/myapplication/login/ForgetPass2.kt
@@ -5,16 +5,17 @@ import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.widget.Toast
 import com.example.myapplication.R
-import kotlinx.android.synthetic.main.forgetpass3.*
+import kotlinx.android.synthetic.main.forgetpass2.*
 
-class ForgetPass3 : AppCompatActivity() {
+
+class ForgetPass2 : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.forgetpass3)
-        forget3bttn.setOnClickListener {
+        setContentView(R.layout.forgetpass2)
+        forget2bttn.setOnClickListener {
         //nut thay doi kiem tra pass
-            if (forget3pass.text.toString().length>0 && forget3passcon.text.toString().length>0){
-                if (forget3pass.text.toString() == forget3passcon.text.toString()){
+            if (forget2pass.text.toString().length>0 && forget2pass.text.toString().length>0){
+                if (forget2pass.text.toString() == forget2passcon.text.toString()){
                     //Hai password match
                     startActivity(Intent(this, MainActivity::class.java))
                 }else{
