+++ app/src/main/java/com/example/myapplication/model/Teacher.kt
@@ -1,4 +1,4 @@
-package com.example.myapplication
+package com.example.myapplication.model
 
 import javax.security.auth.Subject
