+++ app/src/main/java/com/example/myapplication/model/User.kt
@@ -1,4 +1,4 @@
-package com.example.myapplication
+package com.example.myapplication.model
 
 class User {
     var id : Int=0
