+++ app/src/main/res/layouts/LOGIN/layout/login3.xml
@@ -29,12 +29,14 @@
             android:textSize="20sp" />
 
         <EditText
+            android:inputType="text"
             android:layout_marginTop="20dp"
             android:layout_width="300dp"
             android:layout_height="wrap_content"
             android:hint="@string/username"
             android:layout_gravity="center" />
         <EditText
+            android:inputType="textPassword"
             android:layout_marginTop="10dp"
             android:layout_width="300dp"
             android:layout_height="wrap_content"
