+++ app/src/main/res/layouts/LOGIN/layout/signupsv.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent">
-    tools:context=".SingUp">
+    tools:context=".MainActivity">
     <LinearLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
@@ -28,7 +28,7 @@
             android:textColor="@color/black"
             android:textSize="20sp" />
         <EditText
-            android:inputType="text"
+            android:inputType="textPersonName"
             android:id="@+id/SULname"
             android:layout_width="300dp"
             android:layout_height="wrap_content"
@@ -37,7 +37,7 @@
             android:hint="@string/ho"
             android:padding="10dp" />
         <EditText
-            android:inputType="text"
+            android:inputType="textPersonName"
             android:id="@+id/SUFname"
             android:layout_width="300dp"
             android:layout_height="wrap_content"
